Added tests and changed to unit tests
